fix(alembic): shorten gpu_power_multi_worker revision id to fit varchar(32)
The previous revision id ``2026_05_03_gpu_power_multi_worker`` was 34
characters; alembic_version.version_num is varchar(32) on PostgreSQL,
so the closing UPDATE failed with StringDataRightTruncation:

    UPDATE alembic_version SET version_num='2026_05_03_gpu_power_multi_worker'
    WHERE alembic_version.version_num = '2026_05_02_power_multi_worker'
    -> psycopg2.errors.StringDataRightTruncation

The CPU-branch revision (30 chars) sneaked under the limit. The deploy
of v1.31.4 hit the limit, alembic auto-rolled the migration back, and
ci-deploy reverted git to v1.31.3.

Rename the revision to ``gpu_power_mw_2026_05_03`` (23 chars) and
make the upgrade idempotent so a redeploy succeeds even if the three
GPU power tables were already created by the failed attempt before
the UPDATE crashed (the create_table calls run inside the same
transaction as the version UPDATE on PostgreSQL, so they should have
been rolled back too — but the idempotent guard removes any doubt).
